refactor(engine): split database/tests.rs into themed submodules
`crates/engine/src/database/tests.rs` had grown to 5132 lines across the
Tranche 3 durability and T3-E6 shutdown-hardening work, making it the
hardest file in the engine crate to scan and a recurring merge-conflict
hotspot. Move the test body into `database/tests/` with themed
submodules and keep shared helpers in `tests/mod.rs`.

No test logic changes. Same 116 `#[test]` functions before and after
(verified by `cargo test --list` diff). `cargo test -p strata-engine
--lib` stable at 931 passed / 0 failed / 1 ignored across three
consecutive runs. Integration + executor + clippy + fmt all clean.

Layout:

    tests/
      mod.rs          197 lines   shared helpers + `mod X;` declarations
      open.rs         440 lines   open/close/reopen, WAL recovery,
                                  durability modes, singleton registry,
                                  corrupt-WAL recovery failure
      transactions.rs 167 lines   transaction API tests
      checkpoint.rs   737 lines   checkpoint, compact, gc,
                                  T3-E5 snapshot restart, retention
      shutdown.rs    2046 lines   basic/coordinated shutdown, flush
                                  thread, durability mode, WAL halt/
                                  resume, open_runtime lifecycle, all
                                  T3-E6 acceptance tests
      contention.rs   211 lines   put_direct contention, concurrent GC
      regressions.rs 1341 lines   issue-numbered regressions + the
                                  checkpoint-data-preserves-* suite

Biggest file drops from 5132 → 2046 lines.

Shared helpers kept in `tests/mod.rs`:
  - `impl Database { pub fn open_with_durability(...) }`
  - `pub fn write_wal_txn(...)`
  - `pub fn wait_until(...)`
  - `pub fn create_test_namespace(...)`
  - `pub fn blind_write(...)`
  - `pub struct TestRuntimeSubsystem` + `impl Subsystem for ...`

Section-local helpers moved with their tests:
  - `fn corrupt_wal_segment(...)` → `open.rs`
  - `FreezeCountingSubsystem`, `FlakyFreezeSubsystem` → `shutdown.rs`

Module-mechanics notes:

  - `tests/mod.rs` re-exports (`pub use`) the external `use` lines that
    tests currently rely on (`strata_core`, `strata_durability`,
    `tempfile`, etc.) so each submodule just does `use super::*;` and
    picks everything up without any rewrite in test bodies.
  - `super::test_hooks` is a private sibling module of
    `database::tests`; a non-`pub use super::test_hooks;` in `mod.rs`
    still makes it reachable from submodules because child modules can
    see private items of their parent.
  - `test_source_contains_single_flush_thread_implementation` used
    `include_str!("open.rs")` and `include_str!("mod.rs")` relative to
    the old `tests.rs` (which resolved to the production sources).
    After the move those paths pointed at the test submodules of the
    same name, so they were rewritten to `include_str!("../open.rs")`
    and `include_str!("../mod.rs")`.
